requirements.txt
matplotlib==3.5.1
numpy==1.22.4
opencv_python==4.5.5.62
Pillow==9.4.0
plotly==5.13.0
scipy==1.7.3
setuptools==58.0.4
torch==1.10.1
torchvision==0.11.2
tqdm==4.62.3
